move tlsmanager, config, httpclient and kubernetesclient to utils for go and java
