--- pipelinereposerver/src/main/java/com/greenops/pipelinereposerver/kubernetesclient/KubernetesClient.java
@@ -1,15 +1,0 @@
-package com.greenops.pipelinereposerver.kubernetesclient;
-
-import com.greenops.util.datamodel.git.GitCred;
-import io.kubernetes.client.models.V1Secret;
-
-public interface KubernetesClient {
-
-    boolean storeGitCred(GitCred gitCred, String name);
-
-    GitCred fetchGitCred(String name);
-
-    V1Secret fetchSecretData(String name, String namespace);
-
-    void watchSecretData(String name, String namespace, WatchSecretHandler handler);
-}
