--- pipelinereposerver/src/main/java/com/greenops/pipelinereposerver/kubernetesclient/WatchSecretHandler.java
@@ -1,7 +1,0 @@
-package com.greenops.pipelinereposerver.kubernetesclient;
-
-import io.kubernetes.client.models.V1Secret;
-
-public interface WatchSecretHandler {
-    void handle(V1Secret secret);
-}
