--- pipelinereposerver/src/main/java/com/greenops/pipelinereposerver/tslmanager/TLSManager.java
@@ -1,9 +1,0 @@
-package com.greenops.pipelinereposerver.tslmanager;
-
-import org.apache.tomcat.util.net.SSLHostConfig;
-
-public interface TLSManager {
-    SSLHostConfig getSSLHostConfig() throws Exception;
-
-    void watchHostSSLConfig();
-}
