+++ util/src/main/java/com/greenops/util/kubernetesclient/WatchSecretHandler.java
@@ -1,0 +1,7 @@
+package com.greenops.util.kubernetesclient;
+
+import io.kubernetes.client.models.V1Secret;
+
+public interface WatchSecretHandler {
+    void handle(V1Secret secret);
+}
