+++ util/src/main/java/com/greenops/util/tslmanager/TLSManager.java
@@ -1,0 +1,9 @@
+package com.greenops.util.tslmanager;
+
+import org.apache.tomcat.util.net.SSLHostConfig;
+
+public interface TLSManager {
+    SSLHostConfig getSSLHostConfig(ClientName serverName) throws Exception;
+
+    void watchHostSSLConfig(ClientName serverName);
+}
